+++ index.js
@@ -1,109 +1,104 @@
-var defaultIsMergeableObject = require('is-mergeable-object')
+const isPlainObj = require(`is-plain-obj`)
 
-function emptyTarget(val) {
-	return Array.isArray(val) ? [] : {}
-}
+const defaultIsMergeable = value => Array.isArray(value) || isPlainObj(value)
+const emptyTarget = value => Array.isArray(value) ? [] : {}
 
-function cloneUnlessOtherwiseSpecified(value, options) {
-	return (options.clone !== false && options.isMergeableObject(value))
+const cloneUnlessOtherwiseSpecified = (value, options) => {
+	return (options.clone !== false && options.isMergeable(value))
 		? deepmerge(emptyTarget(value), value, options)
 		: value
 }
 
-function defaultArrayMerge(target, source, options) {
-	return target.concat(source).map(function(element) {
-		return cloneUnlessOtherwiseSpecified(element, options)
-	})
+const defaultArrayMerge = (target, source, options) => {
+	return target.concat(source)
+		.map(element => cloneUnlessOtherwiseSpecified(element, options))
 }
 
-function getMergeFunction(key, options) {
+const getMergeFunction = (key, options) => {
 	if (!options.customMerge) {
 		return deepmerge
 	}
-	var customMerge = options.customMerge(key)
-	return typeof customMerge === 'function' ? customMerge : deepmerge
+
+	const customMerge = options.customMerge(key)
+	return typeof customMerge === `function` ? customMerge : deepmerge
 }
 
-function getEnumerableOwnPropertySymbols(target) {
+const getEnumerableOwnPropertySymbols = target => {
 	return Object.getOwnPropertySymbols
-		? Object.getOwnPropertySymbols(target).filter(function(symbol) {
-			return target.propertyIsEnumerable(symbol)
-		})
+		? Object.getOwnPropertySymbols(target)
+			.filter(symbol => target.propertyIsEnumerable(symbol))
 		: []
 }
 
-function getKeys(target) {
-	return Object.keys(target).concat(getEnumerableOwnPropertySymbols(target))
-}
+const getKeys = target => Object.keys(target).concat(getEnumerableOwnPropertySymbols(target))
 
-function propertyIsOnObject(object, property) {
+const propertyIsOnObject = (object, property) => {
 	try {
 		return property in object
-	} catch(_) {
+	} catch (_) {
 		return false
 	}
 }
 
 // Protects from prototype poisoning and unexpected merging up the prototype chain.
-function propertyIsUnsafe(target, key) {
+const propertyIsUnsafe = (target, key) => {
 	return propertyIsOnObject(target, key) // Properties are safe to merge if they don't exist in the target yet,
 		&& !(Object.hasOwnProperty.call(target, key) // unsafe if they exist up the prototype chain,
 			&& Object.propertyIsEnumerable.call(target, key)) // and also unsafe if they're nonenumerable.
 }
 
-function mergeObject(target, source, options) {
-	var destination = {}
-	if (options.isMergeableObject(target)) {
-		getKeys(target).forEach(function(key) {
-			destination[key] = cloneUnlessOtherwiseSpecified(target[key], options)
-		})
+const mergeObject = (target, source, options) => {
+	const destination = {}
+
+	if (options.isMergeable(target)) {
+		getKeys(target)
+			.forEach(key => destination[key] = cloneUnlessOtherwiseSpecified(target[key], options))
 	}
-	getKeys(source).forEach(function(key) {
+
+	getKeys(source).forEach(key => {
 		if (propertyIsUnsafe(target, key)) {
 			return
 		}
 
-		if (!options.isMergeableObject(source[key]) || !propertyIsOnObject(target, key)) {
-			destination[key] = cloneUnlessOtherwiseSpecified(source[key], options)
-		} else {
+		if (propertyIsOnObject(target, key) && options.isMergeable(source[key])) {
 			destination[key] = getMergeFunction(key, options)(target[key], source[key], options)
+		} else {
+			destination[key] = cloneUnlessOtherwiseSpecified(source[key], options)
 		}
 	})
+
 	return destination
 }
 
 function deepmerge(target, source, opts) {
-	var options = {
+	const options = {
 		clone: opts && opts.clone,
 		customMerge: opts && opts.customMerge,
 		arrayMerge: (opts && opts.arrayMerge) || defaultArrayMerge,
-		isMergeableObject: (opts && opts.isMergeableObject) || defaultIsMergeableObject,
+		isMergeable: (opts && opts.isMergeable) || defaultIsMergeable,
 		// cloneUnlessOtherwiseSpecified is added to `options` so that custom arrayMerge()
 		// implementations can use it. The caller may not replace it.
 		cloneUnlessOtherwiseSpecified: cloneUnlessOtherwiseSpecified
 	}
 
-	var sourceIsArray = Array.isArray(source)
-	var targetIsArray = Array.isArray(target)
-	var sourceAndTargetTypesMatch = sourceIsArray === targetIsArray
+	const sourceIsArray = Array.isArray(source)
+	const targetIsArray = Array.isArray(target)
+	const sourceAndTargetTypesMatch = sourceIsArray === targetIsArray
 
 	if (!sourceAndTargetTypesMatch) {
 		return cloneUnlessOtherwiseSpecified(source, options)
 	} else if (sourceIsArray) {
 		return options.arrayMerge(target, source, options)
-	} else {
-		return mergeObject(target, source, options)
 	}
+	return mergeObject(target, source, options)
 }
 
-deepmerge.all = function deepmergeAll(array, options) {
+deepmerge.all = (array, options) => {
 	if (!Array.isArray(array)) {
-		throw new Error('first argument should be an array')
+		throw new Error(`first argument should be an array`)
 	}
 
-	return array.reduce(function(prev, next) {
-		return deepmerge(prev, next, options)
-	}, {})
+	return array.reduce((prev, next) => deepmerge(prev, next, options), {})
 }
 
 module.exports = deepmerge